voice_agent/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
twilio>=9.0.0
google-generativeai>=0.8.0
